Add files via upload
Google Play Store team is about to launch a new feature where in certain apps that are promising are boosted in visibility.
The task is to understand what makes an app perform well - size? price? category? we have to analyze and find some insights.the final output of the analysis would be presented to business as insights with supporting data/visualizations.